requirements.txt
aiogram==2.23.1
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
certifi==2024.2.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
fasteners==0.19
frozenlist==1.4.1
humanfriendly==10.0
idna==3.6
magic-filter==1.0.12
multidict==6.0.5
property-manager==3.0
pydantic==2.5.3
pyreadline3==3.4.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
SpeechRecognition==3.9.0
typing_extensions==4.10.0
urllib3==2.2.1
verboselogs==1.7
yarl==1.9.4
